Update src/koans/13_recursion.clj
Made a separate function for the factorial function applied to bigints.

The reason for this is that after successfully implementing a solution for the "Simple things may appear simple." koan and the other non-bigint factorial koans, the `script/run` application throws a java.lang.StackOverflowError.  Instead, it should signal to the user that he's working on the next koan.