Update 05_maps.clj
"Often you will need to get the keys, but the order is undependable"
The keys value pairs are already ordered by key. Why not make the original map un-ordered to force the user to think about sorting them?

"You can get the values in a similar way"
Two of the three values are already provided in the answer, the answer to this challenge can be arrived at by providing the missing one through process of elimination without thinking about sorting.